fix(frontend): use translated server errors for ServerErrorContent component
