+++ frontend/src/components/form/ServerErrorContent.vue
@@ -1,41 +1,27 @@
 <template>
-  <div v-if="!!serverError.name || !!serverError.response">
-    <template
-      v-if="
-        serverError.name === 'ServerException' &&
-        serverError.response &&
-        serverError.response.status === 422
-      "
-    >
-      <div class="title">Validation error</div>
-      <ul>
-        <li
-          v-for="(violation, index) in serverError.response.data.violations"
-          :key="index"
-        >
-          <div>
-            <b>{{ violation.propertyPath }}</b
-            >: {{ violation.message }}
-          </div>
-        </li>
-      </ul>
-    </template>
-    <template v-else>
-      {{ serverError.message }}
-    </template>
-  </div>
-  <div v-else>
-    {{ serverError }}
+  <div>
+    <ul>
+      <li v-for="(error, index) in errorList" :key="index">
+        {{ error }}
+      </li>
+    </ul>
   </div>
 </template>
 
 <script>
+import { violationsToFlatArray } from '@/helpers/serverError'
+
 export default {
   name: 'ServerErrorContent',
   props: {
     serverError: {
       type: [Object, String, Error],
       default: null,
+    },
+  },
+  computed: {
+    errorList() {
+      return violationsToFlatArray(this.serverError, this.$i18n)
     },
   },
 }
