+++ frontend/src/helpers/serverError.js
@@ -66,4 +66,28 @@ const transformViolations = (error, i18n = null) => {
   return serverErrorMessages
 }
 
-export { serverErrorToString, transformViolations }
+function violationsToFlatArray(e, i18n) {
+  const violationsObject = transformViolations(e, i18n)
+  const violations = Object.entries(violationsObject)
+  if (violations.length === 0) {
+    return []
+  }
+  if (violations.length === 1 && violationsObject[0]) {
+    return [violationsObject[0]]
+  }
+  const toArray = (element) => {
+    if (Array.isArray(element)) {
+      return element
+    }
+    return [element]
+  }
+  const result = []
+  for (const [key, value] of Object.entries(violationsObject)) {
+    for (const message of toArray(value)) {
+      result.push(`${key}: ${message}`)
+    }
+  }
+  return result
+}
+
+export { serverErrorToString, transformViolations, violationsToFlatArray }
